requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# AI and ML
litellm>=1.37.0
openai>=1.0.0

# Bioinformatics
biopython>=1.81

# PDF generation
fpdf2>=2.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
openpyxl>=3.1.0

# Development and testing
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: for enhanced analytics
plotly-express>=0.4.1
networkx>=3.0